Feature/system

+++ Minimo/Assets/02. Scripts/UI/Produce/Advanced/ProduceTaskBtn.cs
@@ -79,7 +79,7 @@ public class ProduceTaskBtn : MonoBehaviour
         else if (_produceTask?.CurrentState is ActiveState)
         {
             var useCashPanel = App.GetManager<UIManager>().GetPanel<UseCashPanel>();
-            var currentStarCount = (_produceTask.RemainTime / 30) + 1;
+            var currentStarCount = (int)(_produceTask.RemainTime / 30) + 1;
             useCashPanel.OpenPanel(UseCashType.Produce, 
                 currentStarCount, 
                 ()=>_produceManager.HarvestEarly());
